Created component for quicktimer + css cleanup.
